+++ examples/example-widget-react-sdk-2.x/src/main.tsx
@@ -14,5 +14,5 @@ createRoot(root).render(
     <FoundryWidget config={MainConfig}>
       <App />
     </FoundryWidget>
-  </Theme>
+  </Theme>,
 );