+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/client.test.ontology
 
+## 2.1.0-beta.18
+
+### Patch Changes
+
+- @osdk/api@2.1.0-beta.18
+
 ## 2.1.0-beta.17
 
 ### Patch Changes